refactor: Facade 패턴 적용
