+++ backend/src/main/java/com/festago/auth/presentation/AuthController.java
@@ -1,7 +1,7 @@
 package com.festago.auth.presentation;
 
 import com.festago.auth.annotation.Member;
-import com.festago.auth.application.AuthService;
+import com.festago.auth.application.AuthFacadeService;
 import com.festago.auth.dto.LoginRequest;
 import com.festago.auth.dto.LoginResponse;
 import io.swagger.v3.oas.annotations.Operation;
@@ -19,16 +19,16 @@ import org.springframework.web.bind.annotation.RestController;
 @Tag(name = "로그인 관련 요청")
 public class AuthController {
 
-    private final AuthService authService;
+    private final AuthFacadeService authFacadeService;
 
-    public AuthController(AuthService authService) {
-        this.authService = authService;
+    public AuthController(AuthFacadeService authFacadeService) {
+        this.authFacadeService = authFacadeService;
     }
 
     @PostMapping("/oauth2")
     @Operation(description = "소셜 엑세스 토큰을 기반으로 로그인 요청을 보낸다.", summary = "OAuth2 로그인")
     public ResponseEntity<LoginResponse> login(@RequestBody LoginRequest request) {
-        LoginResponse response = authService.login(request);
+        LoginResponse response = authFacadeService.login(request);
         return ResponseEntity.ok()
             .body(response);
     }
@@ -37,7 +37,7 @@ public class AuthController {
     @SecurityRequirement(name = "bearerAuth")
     @Operation(description = "회원 탈퇴 요청을 보낸다.", summary = "유저 회원 탈퇴")
     public ResponseEntity<Void> deleteMember(@Member Long memberId) {
-        authService.deleteMember(memberId);
+        authFacadeService.deleteMember(memberId);
         return ResponseEntity.ok()
             .build();
     }
